refactor: updatePrevPosition 함수 별도 유틸 파일로 분리
