+++ frontend/src/components/play/Character.tsx
@@ -6,6 +6,7 @@ import { saveCharacterInfo } from 'utils/saveCharacterInfo';
 import { createShadowImgAndTrackTouch } from 'utils/createShadowImgAndTrackTouch';
 import { updateShadowImgAndTrackTouch } from 'utils/updateShadowImgAndTrackTouch';
 import { removeShadowImgAndDispatchEndEvent } from 'utils/removeShadowImgAndDispatchEndEvent';
+import { updatePrevPosition } from 'utils/updatePrevPosition';
 import { usePlayStore } from 'stores/playStore';
 
 interface CharacterProps {
@@ -16,32 +17,15 @@ const Character = ({ characterInfo }: CharacterProps) => {
   const { characterOption, characterSize, characterKey } = characterInfo;
   const imageSrc = CHARACTER_MAP[characterOption][characterSize];
   const dragShadowImgRef = useRef<HTMLImageElement | null>(null); // 드래그 시 생성되는 쉐도우 이미지 참조
-
   const { selectedCharacterKey, setSelectedCharacterKey, setPrevPosition } =
     usePlayStore();
-
   const isSelected = selectedCharacterKey === characterKey;
-
-  const updatePrevPosition = (target: HTMLElement) => {
-    const parentSquare = target.closest('.square');
-
-    if (parentSquare) {
-      const rowMatch = parentSquare.className.match(/row-(\d+)/);
-      const colMatch = parentSquare.className.match(/col-(\d+)/);
-
-      if (rowMatch && colMatch) {
-        const row = parseInt(rowMatch[1], 10);
-        const col = parseInt(colMatch[1], 10);
-        setPrevPosition({ row, col });
-      }
-    }
-  };
 
   // 모바일 환경
   const handleTouchStart = (e: React.TouchEvent<HTMLImageElement>) => {
     createShadowImgAndTrackTouch(e, characterInfo, imageSrc, dragShadowImgRef);
     setSelectedCharacterKey(characterKey);
-    updatePrevPosition(e.currentTarget);
+    updatePrevPosition(e.currentTarget, setPrevPosition);
   };
 
   const handleTouchMove = (e: React.TouchEvent<HTMLImageElement>) => {
@@ -58,7 +42,7 @@ const Character = ({ characterInfo }: CharacterProps) => {
   const handleDragStart = (e: React.DragEvent<HTMLImageElement>) => {
     saveCharacterInfo(e, characterInfo);
     setSelectedCharacterKey(characterKey);
-    updatePrevPosition(e.currentTarget);
+    updatePrevPosition(e.currentTarget, setPrevPosition);
   };
 
   return (
